requirements.txt
cerberus
cryptography
dkimpy
dnspython
jinja2
python-gnupg
pyyaml

# Testing
pytest
pylint
tox
aiosmtpd

# Distribution
twine
wheel
